Update for Visual Studio 2022 development
Includes using the automatic migration to the current (as of this
commit) version of Visual Studio, adding a standard ignore file to
reduce the amount of excess files included in the repository, and
renaming a problematic define in glh that caused compilation errors in
the system_error and xios libraries.

__NOTE__: In the process of testing a successfully-compiled version of
VirtualScape, the resulting application used the default operating
system window types, creating lag spikes as MFC tried to keep up. I
intend to include this information in the relevant pull request when the
time comes.